First commited as before_9761_rev1

+++ src/main/java/net/spy/memcached/protocol/binary/GetlOperationImpl.java
@@ -31,7 +31,7 @@ import net.spy.memcached.ops.GetlOperation;
 public class GetlOperationImpl extends SingleKeyOperationImpl implements
     GetlOperation {
 
-  static final byte GETL_CMD = (byte) 0x94;
+  static final int GETL_CMD = 0x94;
 
   /**
    * Length of the extra header stuff for a GET response.
